refactor: kafka producer infra - domain 분리
