+++ src/main/java/io/hhplus/concertbook/domain/KafkaProducer/TestProducer.java
@@ -1,0 +1,11 @@
+package io.hhplus.concertbook.domain.KafkaProducer;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.kafka.core.KafkaTemplate;
+import org.springframework.stereotype.Component;
+
+public interface TestProducer {
+
+    public void create();
+
+}
