+++ src/main/java/io/hhplus/concertbook/domain/service/TestKafkaService.java
@@ -1,9 +1,9 @@
 package io.hhplus.concertbook.domain.service;
 
-import io.hhplus.concertbook.infra.KafkaProducer.TestProducer;
+import io.hhplus.concertbook.domain.KafkaProducer.TestProducer;
+import io.hhplus.concertbook.infra.KafkaProducer.TestProducerImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.GetMapping;
 
 @Service
 public class TestKafkaService {
