+++ src/main/java/io/hhplus/concertbook/infra/KafkaProducer/BookProducerImpl.java
@@ -1,5 +1,6 @@
 package io.hhplus.concertbook.infra.KafkaProducer;
 
+import io.hhplus.concertbook.domain.KafkaProducer.BookProducer;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.kafka.core.KafkaTemplate;
@@ -7,11 +8,13 @@ import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class PayProducer {
+public class BookProducerImpl implements BookProducer {
+
     @Autowired
     private KafkaTemplate<String, Object> kafkaTemplate;
 
     public void send(String topic,String key, Long outboxId) {
         kafkaTemplate.send(topic,key,outboxId);
     }
+
 }
