+++ src/main/java/io/hhplus/concertbook/infra/KafkaProducer/PayProducerImpl.java
@@ -1,5 +1,6 @@
 package io.hhplus.concertbook.infra.KafkaProducer;
 
+import io.hhplus.concertbook.domain.KafkaProducer.PayProducer;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.kafka.core.KafkaTemplate;
@@ -7,13 +8,11 @@ import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class BookProducer {
-
+public class PayProducerImpl implements PayProducer {
     @Autowired
     private KafkaTemplate<String, Object> kafkaTemplate;
 
     public void send(String topic,String key, Long outboxId) {
         kafkaTemplate.send(topic,key,outboxId);
     }
-
 }
